refactor: update beneficiary handling and introduce asset conversion utility
- Renamed groupBeneficiariesByAsset to groupBeneficiariesByAssetIndex for clarity and updated its implementation to use a ValidatorService for network validation.
- Introduced a new utility function convertAssetToUSD to handle asset conversion to USD, enhancing the codebase's functionality.
- Updated various components to utilize the new grouping and conversion methods, ensuring consistent handling of beneficiaries and their associated values.
- Removed deprecated methods and cleaned up related code to streamline the overall structure.
