+++ src/app/_shared-components/PostDetails/BeneficiariesDetails/BeneficiariesDetails.tsx
@@ -2,17 +2,49 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { IBeneficiary } from '@/_shared/types';
+import { ENetwork, IBeneficiary, IBeneficiaryInput } from '@/_shared/types';
 import { useTranslations } from 'next-intl';
 import { useMemo, useState } from 'react';
 import { ChevronRightIcon } from 'lucide-react';
-import { groupBeneficiariesByAssetWithAddress } from '@/app/_client-utils/beneficiaryUtils';
 import { getCurrentNetwork } from '@/_shared/_utils/getCurrentNetwork';
 import { cn } from '@/lib/utils';
+import { NETWORKS_DETAILS } from '@/_shared/_constants/networks';
+import { BN } from '@polkadot/util';
+import { ValidatorService } from '@/_shared/_services/validator_service';
 import classes from './BeneficiariesDetails.module.scss';
 import { Button } from '../../Button';
 import BeneficiariesDetailsDialog from './BeneficiariesDetailsDialog';
 import BeneficiaryItem from './BeneficiaryItem';
+
+export const aggregateBeneficiariesByAsset = (beneficiaries: IBeneficiaryInput[] | undefined | null, network: ENetwork): Record<string, { amount: BN; addresses: string[] }> => {
+	if (!beneficiaries || !Array.isArray(beneficiaries) || !ValidatorService.isValidNetwork(network)) {
+		return {};
+	}
+
+	return beneficiaries.reduce((acc: Record<string, { amount: BN; addresses: string[] }>, curr: IBeneficiaryInput) => {
+		if (!curr) return acc;
+
+		const assetId = curr.assetId || NETWORKS_DETAILS[network as ENetwork].tokenSymbol;
+
+		if (!assetId) return acc;
+
+		if (!acc[assetId as string]) {
+			acc[assetId as string] = { amount: new BN(0), addresses: [] };
+		}
+
+		try {
+			const amount = new BN(curr.amount || '0');
+			acc[assetId as string] = {
+				amount: acc[assetId as string].amount.add(amount),
+				addresses: [...new Set([...acc[assetId as string].addresses, curr.address || ''])]
+			};
+		} catch (error) {
+			console.error(`Error processing beneficiary amount: ${error}`);
+		}
+
+		return acc;
+	}, {});
+};
 
 // Main component
 function BeneficiariesDetails({ beneficiaries }: { beneficiaries: IBeneficiary[] }) {
@@ -20,7 +52,7 @@ function BeneficiariesDetails({ beneficiaries }: { beneficiaries: IBeneficiary[]
 	const [openDialog, setOpenDialog] = useState(false);
 	const network = getCurrentNetwork();
 
-	const groupedBeneficiaries = useMemo(() => groupBeneficiariesByAssetWithAddress(beneficiaries, network), [beneficiaries, network]);
+	const groupedBeneficiaries = useMemo(() => aggregateBeneficiariesByAsset(beneficiaries, network), [beneficiaries, network]);
 
 	// Early return if no beneficiaries
 	if (!beneficiaries?.length) {
